Fixed the email when user is registering (to the site and the newsletter simultaneously)

+++ ps_emailsubscription.php
@@ -45,6 +45,8 @@ class Ps_Emailsubscription extends Module implements WidgetInterface
 
     const LEGAL_PRIVACY = 'LEGAL_PRIVACY';
 
+    protected $_origin_newsletter;
+
     const TPL_COLUMN = 'ps_emailsubscription-column.tpl';
     const TPL_DEFAULT = 'ps_emailsubscription.tpl';
 
@@ -68,7 +70,7 @@ class Ps_Emailsubscription extends Module implements WidgetInterface
 
         $this->entity_manager = $entity_manager;
 
-        $this->version = '2.5.2';
+        $this->version = '2.5.3';
         $this->author = 'PrestaShop';
         $this->error = false;
         $this->valid = false;
@@ -106,11 +108,13 @@ class Ps_Emailsubscription extends Module implements WidgetInterface
                     'actionFrontControllerSetMedia',
                     'displayFooterBefore',
                     'actionCustomerAccountAdd',
+                    'actionObjectCustomerUpdateBefore',
                     'additionalCustomerFormFields',
                     'displayAdminCustomersForm',
                     'registerGDPRConsent',
                     'actionDeleteGDPRCustomer',
                     'actionExportGDPRData',
+                    'actionCustomerAccountUpdate',
                 )
             )
         ) {
@@ -885,14 +889,51 @@ class Ps_Emailsubscription extends Module implements WidgetInterface
     {
         //if e-mail of the created user address has already been added to the newsletter through the ps_emailsubscription module,
         //we delete it from ps_emailsubscription table to prevent duplicates
+        if (empty($params['newCustomer'])) {
+            return;
+        }
         $id_shop = $params['newCustomer']->id_shop;
         $email = $params['newCustomer']->email;
+        $newsletter = $params['newCustomer']->newsletter;
         if (Validate::isEmail($email)) {
             if ($params['newCustomer']->newsletter && $code = Configuration::get('NW_VOUCHER_CODE')) {
                 $this->sendVoucher($email, $code);
             }
 
             return (bool) Db::getInstance()->execute('DELETE FROM ' . _DB_PREFIX_ . 'emailsubscription WHERE id_shop=' . (int) $id_shop . ' AND email=\'' . pSQL($email) . "'");
+        }
+
+        if ($newsletter) {
+            if (Configuration::get('NW_CONFIRMATION_EMAIL')) {// send confirmation email
+                $this->sendConfirmationEmail($params['newCustomer']->email);
+            }
+            if ($code = Configuration::get('NW_VOUCHER_CODE')) {// send voucher
+                $this->sendVoucher($params['newCustomer']->email, $code);
+            }
+        }
+
+        return true;
+    }
+
+   public function hookActionObjectCustomerUpdateBefore($params)
+   {
+       $customer = new Customer($params['object']->id);
+       $this->_origin_newsletter = (int)$customer->newsletter;
+   }
+
+    public function hookActionCustomerAccountUpdate($params)
+    {
+        if ($this->_origin_newsletter || !$params['customer']->newsletter) {
+            return;
+        }
+        if (Configuration::get('NW_CONFIRMATION_EMAIL')) {// send confirmation email
+            $this->sendConfirmationEmail($params['customer']->email);
+        }
+        if ($code = Configuration::get('NW_VOUCHER_CODE')) {
+            $cartRule = CartRuleCore::getCartsRuleByCode($code, Context::getContext()->language->id);
+            if (! Order::getDiscountsCustomer($params['customer']->id, $cartRule[0])) {// send voucher
+                $this->sendVoucher($params['customer']->email, $code);
+            }
         }
 
         return true;
